+++ configs.example/datamodule/streamed_samples.yaml
@@ -1,4 +1,4 @@
-_target_: pvnet.data.UKRegionalDataModule
+_target_: pvnet.datamodule.UKRegionalDataModule
 # Path to the data configuration yaml file. You can find examples in the configuration subdirectory
 # in configs.example/datamodule/configuration
 # Use the full local path such as: /FULL/PATH/PVNet/configs/datamodule/configuration/example_configuration.yaml"
