--- pvnet/data/__init__.py
@@ -1,3 +1,0 @@
-"""Data parts"""
-from .site_datamodule import SitesDataModule
-from .uk_regional_datamodule import UKRegionalDataModule
